+++ test/vanilla/Expected/AcceptanceTests/BodyArray/bodyarray/models/__init__.py
@@ -10,17 +10,17 @@
 # --------------------------------------------------------------------------
 
 try:
-    from .product_py3 import Product
-    from .error_py3 import Error, ErrorException
+    from ._models_py3 import Error, ErrorException
+    from ._models_py3 import Product
 except (SyntaxError, ImportError):
-    from .product import Product
-    from .error import Error, ErrorException
-from .auto_rest_swagger_bat_array_service_enums import (
+    from ._models import Error, ErrorException
+    from ._models import Product
+from ._auto_rest_swagger_bat_array_service_enums import (
     FooEnum,
 )
 
 __all__ = [
-    'Product',
     'Error', 'ErrorException',
+    'Product',
     'FooEnum',
 ]