+++ test/vanilla/Expected/AcceptanceTests/ExtensibleEnums/extensibleenumsswagger/models/__init__.py
@@ -10,10 +10,10 @@
 # --------------------------------------------------------------------------
 
 try:
-    from .pet_py3 import Pet
+    from ._models_py3 import Pet
 except (SyntaxError, ImportError):
-    from .pet import Pet
-from .pet_store_inc_enums import (
+    from ._models import Pet
+from ._pet_store_inc_enums import (
     DaysOfWeekExtensibleEnum,
     IntEnum,
 )