+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # Change Log
 
+
+### 2019-02-13 - 4.0.67
+
+- All models will now be generated in two files `_models` and `_models_py3`, and paging in a page file
+- Breaking changes: Import of models and enums from package must be done from package.models
+
+### 2019-02-11 - 4.0.66
+
+- Fix async cross-link documentation
+
 ### 2019-02-08 - 4.0.65
 
 - New version of async generation. Requires msrest 0.6.3 and msrestazure 0.6.0.